+++ cmd/main_test.go
@@ -16,15 +16,15 @@ package cmd
 import (
 	"encoding/base64"
 	"io"
-	"io/ioutil"
 	"net/http"
 	"net/http/httptest"
+	"os"
 	"strings"
 	"testing"
 
-	"github.com/go-kit/log"
 	"github.com/prometheus-community/json_exporter/config"
 	pconfig "github.com/prometheus/common/config"
+	"github.com/prometheus/common/promslog"
 )
 
 func TestFailIfSelfSignedCA(t *testing.T) {
@@ -34,10 +34,10 @@ func TestFailIfSelfSignedCA(t *testing.T) {
 
 	req := httptest.NewRequest("GET", "http://example.com/foo"+"?module=default&target="+target.URL, nil)
 	recorder := httptest.NewRecorder()
-	probeHandler(recorder, req, log.NewNopLogger(), config.Config{Modules: map[string]config.Module{"default": {}}})
+	probeHandler(recorder, req, promslog.NewNopLogger(), config.Config{Modules: map[string]config.Module{"default": {}}})
 
 	resp := recorder.Result()
-	body, _ := ioutil.ReadAll(resp.Body)
+	body, _ := io.ReadAll(resp.Body)
 
 	if resp.StatusCode != http.StatusServiceUnavailable {
 		t.Fatalf("Fail if (not strict) selfsigned CA test fails unexpectedly, got %s", body)
@@ -61,10 +61,10 @@ func TestSucceedIfSelfSignedCA(t *testing.T) {
 
 	req := httptest.NewRequest("GET", "http://example.com/foo"+"?module=default&target="+target.URL, nil)
 	recorder := httptest.NewRecorder()
-	probeHandler(recorder, req, log.NewNopLogger(), c)
+	probeHandler(recorder, req, promslog.NewNopLogger(), c)
 
 	resp := recorder.Result()
-	body, _ := ioutil.ReadAll(resp.Body)
+	body, _ := io.ReadAll(resp.Body)
 
 	if resp.StatusCode != http.StatusOK {
 		t.Fatalf("Succeed if (not strict) selfsigned CA test fails unexpectedly, got %s", body)
@@ -78,7 +78,7 @@ func TestDefaultModule(t *testing.T) {
 
 	req := httptest.NewRequest("GET", "http://example.com/foo"+"?target="+target.URL, nil)
 	recorder := httptest.NewRecorder()
-	probeHandler(recorder, req, log.NewNopLogger(), config.Config{Modules: map[string]config.Module{"default": {}}})
+	probeHandler(recorder, req, promslog.NewNopLogger(), config.Config{Modules: map[string]config.Module{"default": {}}})
 
 	resp := recorder.Result()
 	if resp.StatusCode != http.StatusOK {
@@ -87,7 +87,7 @@ func TestDefaultModule(t *testing.T) {
 
 	// Module doesn't exist.
 	recorder = httptest.NewRecorder()
-	probeHandler(recorder, req, log.NewNopLogger(), config.Config{Modules: map[string]config.Module{"foo": {}}})
+	probeHandler(recorder, req, promslog.NewNopLogger(), config.Config{Modules: map[string]config.Module{"foo": {}}})
 	resp = recorder.Result()
 	if resp.StatusCode != http.StatusBadRequest {
 		t.Fatalf("Default module test fails unexpectedly, expected 400, got %d", resp.StatusCode)
@@ -97,10 +97,10 @@ func TestDefaultModule(t *testing.T) {
 func TestFailIfTargetMissing(t *testing.T) {
 	req := httptest.NewRequest("GET", "http://example.com/foo", nil)
 	recorder := httptest.NewRecorder()
-	probeHandler(recorder, req, log.NewNopLogger(), config.Config{})
+	probeHandler(recorder, req, promslog.NewNopLogger(), config.Config{})
 
 	resp := recorder.Result()
-	body, _ := ioutil.ReadAll(resp.Body)
+	body, _ := io.ReadAll(resp.Body)
 
 	if resp.StatusCode != http.StatusBadRequest {
 		t.Fatalf("Fail if 'target' query parameter missing test fails unexpectedly, got %s", body)
@@ -119,10 +119,10 @@ func TestDefaultAcceptHeader(t *testing.T) {
 
 	req := httptest.NewRequest("GET", "http://example.com/foo"+"?module=default&target="+target.URL, nil)
 	recorder := httptest.NewRecorder()
-	probeHandler(recorder, req, log.NewNopLogger(), config.Config{Modules: map[string]config.Module{"default": {}}})
+	probeHandler(recorder, req, promslog.NewNopLogger(), config.Config{Modules: map[string]config.Module{"default": {}}})
 
 	resp := recorder.Result()
-	body, _ := ioutil.ReadAll(resp.Body)
+	body, _ := io.ReadAll(resp.Body)
 
 	if resp.StatusCode != http.StatusOK {
 		t.Fatalf("Default 'Accept: application/json' header test fails unexpectedly, got %s", body)
@@ -151,12 +151,12 @@ func TestCorrectResponse(t *testing.T) {
 
 		req := httptest.NewRequest("GET", "http://example.com/foo"+"?module=default&target="+target.URL+test.ServeFile, nil)
 		recorder := httptest.NewRecorder()
-		probeHandler(recorder, req, log.NewNopLogger(), c)
+		probeHandler(recorder, req, promslog.NewNopLogger(), c)
 
 		resp := recorder.Result()
-		body, _ := ioutil.ReadAll(resp.Body)
+		body, _ := io.ReadAll(resp.Body)
 
-		expected, _ := ioutil.ReadFile(test.ResponseFile)
+		expected, _ := os.ReadFile(test.ResponseFile)
 
 		if test.ShouldSucceed && string(body) != string(expected) {
 			t.Fatalf("Correct response validation test %d fails unexpectedly.\nGOT:\n%s\nEXPECTED:\n%s", i, body, expected)
@@ -191,10 +191,10 @@ func TestBasicAuth(t *testing.T) {
 		},
 	}
 
-	probeHandler(recorder, req, log.NewNopLogger(), c)
+	probeHandler(recorder, req, promslog.NewNopLogger(), c)
 
 	resp := recorder.Result()
-	body, _ := ioutil.ReadAll(resp.Body)
+	body, _ := io.ReadAll(resp.Body)
 
 	if resp.StatusCode != http.StatusOK {
 		t.Fatalf("BasicAuth test fails unexpectedly. Got: %s", body)
@@ -222,10 +222,10 @@ func TestBearerToken(t *testing.T) {
 		}},
 	}
 
-	probeHandler(recorder, req, log.NewNopLogger(), c)
+	probeHandler(recorder, req, promslog.NewNopLogger(), c)
 
 	resp := recorder.Result()
-	body, _ := ioutil.ReadAll(resp.Body)
+	body, _ := io.ReadAll(resp.Body)
 
 	if resp.StatusCode != http.StatusOK {
 		t.Fatalf("BearerToken test fails unexpectedly. Got: %s", body)
@@ -258,10 +258,10 @@ func TestHTTPHeaders(t *testing.T) {
 		},
 	}
 
-	probeHandler(recorder, req, log.NewNopLogger(), c)
+	probeHandler(recorder, req, promslog.NewNopLogger(), c)
 
 	resp := recorder.Result()
-	body, _ := ioutil.ReadAll(resp.Body)
+	body, _ := io.ReadAll(resp.Body)
 
 	if resp.StatusCode != http.StatusOK {
 		t.Fatalf("Setting custom headers failed unexpectedly. Got: %s", body)
@@ -321,10 +321,10 @@ func TestBodyPostTemplate(t *testing.T) {
 			},
 		}
 
-		probeHandler(recorder, req, log.NewNopLogger(), c)
+		probeHandler(recorder, req, promslog.NewNopLogger(), c)
 
 		resp := recorder.Result()
-		respBody, _ := ioutil.ReadAll(resp.Body)
+		respBody, _ := io.ReadAll(resp.Body)
 
 		if resp.StatusCode != http.StatusOK {
 			t.Fatalf("POST body content failed. Got: %s", respBody)
@@ -420,10 +420,10 @@ func TestBodyPostQuery(t *testing.T) {
 			},
 		}
 
-		probeHandler(recorder, req, log.NewNopLogger(), c)
+		probeHandler(recorder, req, promslog.NewNopLogger(), c)
 
 		resp := recorder.Result()
-		respBody, _ := ioutil.ReadAll(resp.Body)
+		respBody, _ := io.ReadAll(resp.Body)
 
 		if resp.StatusCode != http.StatusOK {
 			t.Fatalf("POST body content failed. Got: %s", respBody)